+++ test.js
@@ -1,0 +1,17 @@
+/* eslint-disable */
+/*
+hum sorry, it's not automated yet
+
+./npmpub.js --verbose --debug --skip-status --skip-fetch --skip-compare --skip-cleanup --dry
+./npmpub.js --verbose --debug --skip-fetch --skip-compare --skip-cleanup --dry
+./npmpub.js --verbose --debug --skip-compare --skip-cleanup --dry
+./npmpub.js --verbose --skip-cleanup --dry
+./npmpub.js --skip-cleanup --dry
+./npmpub.js --dry
+
+Then when everything is ready, pray and try
+
+npm run release
+
+And now you know.
+*/